backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-RESTX==1.3.0  # For Swagger/OpenAPI documentation
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
yt-dlp==2023.11.16
scdl==2.7.3
pyacoustid==1.3.0
numpy==1.24.3  # Required by audfprint
scipy==1.11.3  # Required by audfprint
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.3
python-magic==0.4.27
structlog==23.2.0  # For structured logging
python-json-logger==2.0.7  # JSON formatter for logs
